% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiverseworkshop3-package.R
\docType{package}
\name{epiverseworkshop3-package}
\alias{epiverseworkshop3}
\alias{epiverseworkshop3-package}
\title{epiverseworkshop3: The Title of the Project}
\description{
A paragraph providing a full description of the project (on several lines...)
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/scarabel/epiverseworkshop3}
  \item Report bugs at \url{https://github.com/scarabel/epiverseworkshop3/issues}
}

}
\author{
\strong{Maintainer}: Francesca Scarabel \email{francesca.scarabel@gmail.com} (\href{https://orcid.org/0000-0003-0250-4555}{ORCID}) [copyright holder]

}
\keyword{internal}
